rename -> useMigrate
